+++ client/src/pages/Game/pages/Table/components/EventList/components/EventLine.tsx
@@ -18,6 +18,7 @@ import { memo, useEffect, useMemo, useRef } from 'react'
 import styled, { css, keyframes } from 'styled-components'
 import { CardResourceIcon } from '../../CardResourceIcon/CardResourceIcon'
 import { ResourceIcon } from '../../ResourceIcon/ResourceIcon'
+import { formatTime } from '@/utils/formatTime'
 
 type Props = {
 	event: GameEvent
@@ -75,14 +76,6 @@ const GlobalEventSpan = ({ eventCode }: { eventCode: string }) => {
 			</CardSpanE>
 		</>
 	)
-}
-
-const formatTime = (time: number) => {
-	const hours = Math.floor(time / 3600000)
-	const minutes = Math.floor((time % 3600000) / 60000)
-	const seconds = ((time % 60000) / 1000).toFixed(0)
-
-	return `${hours > 0 ? hours + ':' : ''}${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`
 }
 
 export const EventLine = ({ event, animated, onDone, timestamp }: Props) => {
