+++ client/src/pages/Game/pages/Table/components/InYourFaceEvents/InYourFaceEvents.tsx
@@ -1,18 +1,17 @@
 import { Button, Portal } from '@/components'
 import { ClippedBox } from '@/components/ClippedBox'
 import { Flex } from '@/components/Flex/Flex'
-import { useLocale } from '@/context/LocaleContext'
 import { useAppStore, useToggle } from '@/utils/hooks'
 import { useGameEventsHandler } from '@/utils/useGameEventsHandler'
 import { faBell, faChevronDown } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { EventType, GameEvent } from '@shared/index'
+import { GameEvent } from '@shared/index'
 import { useCallback, useEffect, useState } from 'react'
 import styled, { css, keyframes } from 'styled-components'
 import { InYourFaceEvent } from './components/InYourFaceEvent'
-import { PlayerDidHeader } from './components/PlayerDidHeader'
-import { isInYourFaceEvent } from './utils/isInYourFaceEvent'
 import { InYourFaceEventsList } from './components/InYourFaceEventsList'
+import { InYourFaceEventTitle } from './components/InYourFaceEventTitle'
+import { isInYourFaceEvent } from './utils/isInYourFaceEvent'
 
 export const InYourFaceEvents = () => {
 	const player = useAppStore((state) => state.game.player)
@@ -25,8 +24,6 @@ export const InYourFaceEvents = () => {
 	const [rendered, setRendered] = useState(false)
 	const [events, setEvents] = useState<GameEvent[]>([])
 	const [showList, toggleList] = useToggle()
-
-	const t = useLocale()
 
 	const current = events[0]
 
@@ -49,111 +46,6 @@ export const InYourFaceEvents = () => {
 			setRendered(true)
 		}
 	}, [shown])
-
-	const renderEventHead = useCallback((event: GameEvent) => {
-		switch (event.type) {
-			case EventType.CardPlayed:
-				return (
-					<PlayerDidHeader
-						noSpacing
-						playerId={event.playerId}
-						thing=" played card"
-					/>
-				)
-			case EventType.CardUsed:
-				return (
-					<PlayerDidHeader
-						noSpacing
-						playerId={event.playerId}
-						thing=" used card"
-					/>
-				)
-			case EventType.CompetitionSponsored:
-				return (
-					<PlayerDidHeader
-						noSpacing
-						playerId={event.playerId}
-						thing=" sponsored competition"
-					/>
-				)
-			case EventType.MilestoneBought:
-				return (
-					<PlayerDidHeader
-						noSpacing
-						playerId={event.playerId}
-						thing=" bought milestone"
-					/>
-				)
-			case EventType.ColonyBuilt:
-				return (
-					<PlayerDidHeader
-						noSpacing
-						playerId={event.playerId}
-						thing=" built colony"
-					/>
-				)
-			case EventType.ColonyTrading:
-				return (
-					<PlayerDidHeader
-						noSpacing
-						playerId={event.playerId}
-						thing=" traded with colony"
-					/>
-				)
-			case EventType.StandardProjectBought:
-				return (
-					<PlayerDidHeader
-						playerId={event.playerId}
-						noSpacing
-						thing=" bought standard project"
-					/>
-				)
-			case EventType.StartingSetup:
-				return (
-					<PlayerDidHeader
-						playerId={event.playerId}
-						noSpacing
-						thing=" picked their starting setup"
-					/>
-				)
-			case EventType.TilePlaced:
-				return (
-					<PlayerDidHeader
-						playerId={event.playerId}
-						noSpacing
-						thing=" placed tile"
-					/>
-				)
-			case EventType.ProductionDone:
-				return <CenterText>Production</CenterText>
-			case EventType.MarsTerraformed:
-				return <CenterText>Mars terraformed</CenterText>
-			case EventType.GlobalEventsChanged:
-				return <CenterText>Global events changed</CenterText>
-			case EventType.CurrentGlobalEventExecuted:
-				return <CenterText>Global event executed</CenterText>
-			case EventType.NewGovernment:
-				return <CenterText>New government</CenterText>
-			case EventType.PlayerMovedDelegate:
-				return (
-					<PlayerDidHeader
-						playerId={event.playerId}
-						noSpacing
-						thing={` added delegate to ${t.committeeParties[event.partyCode]}`}
-					/>
-				)
-			case EventType.CommitteePartyActivePolicyActivated:
-				return (
-					<PlayerDidHeader
-						playerId={event.playerId}
-						noSpacing
-						thing=" used ruling policy"
-					/>
-				)
-			default:
-				return null
-		}
-	}, [])
 
 	const handleDismiss = useCallback(() => {
 		setEvents((events) => events.slice(1))
@@ -206,7 +98,7 @@ export const InYourFaceEvents = () => {
 												marginRight: `${(indexReversed + 1) * 0.5}rem`,
 											}}
 										>
-											{renderEventHead(e)}
+											<InYourFaceEventTitle event={e} />
 										</NextEvent>
 									)
 								})}
@@ -331,10 +223,6 @@ const NextEvents = styled.div`
 	position: absolute;
 	margin-top: -1rem;
 	width: 30rem;
-`
-
-const CenterText = styled.div`
-	text-align: center;
 `
 
 const MinimizedButton = styled(Button)`
